shellwords

Function join

pub fn join(args: &[&str]) -> String

Builds a command line string from a list of arguments.

The arguments are combined into a single string with each word separated by a space. Each individual word is escaped as necessary via escape.

Examples

let args = ["There's", "a", "time", "and", "place", "for", "everything"];
assert_eq!(join(&args), "There\\'s a time and place for everything");